// src/enums/task-type.enum.ts
export enum TaskTypeEnum {
  CREATE_PROJECT = 'CREATE_PROJECT',
  GENERATE_COMPONENT = 'GENERATE_COMPONENT',
  ADD_STATE_MANAGEMENT = 'ADD_STATE_MANAGEMENT',
  GENERATE_TESTS = 'GENERATE_TESTS',
  OPTIMIZE_PERFORMANCE = 'OPTIMIZE_PERFORMANCE',
  SETUP_API = 'SETUP_API',
  DEPLOY_PROJECT = 'DEPLOY_PROJECT',
  ANALYZE_REQUIREMENTS = 'ANALYZE_REQUIREMENTS',
  CREATE_ARCHITECTURE = 'CREATE_ARCHITECTURE'
}